Splitting out ObserveGames into a re-usable component
